+++ compiler/main/lambda/transpile_lambdas.h
@@ -1,9 +1,0 @@
-#ifndef TRANSPILE_LAMBDA_H
-#define TRANSPILE_LAMBDA_H
-
-#include "ast/ast.h"
-#include "tables/symtable/symtable.h"
-
-void transpileLambdas(struct AST* ast, struct ST* st);
-
-#endif