+++ parser/main/astnodes/expr/Term.c
@@ -12,7 +12,6 @@
 #include "Expr.h"
 #include "var/Variable.h"
 #include "statements/Call.h"
-#include "Lambda.h"
 
 #include "ast/util/free_ast.h"
 
@@ -55,11 +54,7 @@ struct Term* makeTerm(struct TokenList* tokens) {
 
 	if(tk_kind == LPARENS){
 		
-		if((res->ptr.m11 = makeLambda(copy)) != NULL){
-			res->kind = 11;
-		}else{
-			tryInitExpr(res, copy);
-		}
+		tryInitExpr(res, copy);
 		
 	}else if(tk_kind == STRINGCONST) {