+++ parser/test/astnodes/LambdaTest.h
@@ -1,8 +1,0 @@
-#ifndef LAMBDATEST_H
-#define LAMBDATEST_H
-
-#include <stdbool.h>
-
-void test_lambda();
-
-#endif